fix(quality): clang-tidy modules PART 2 — SAL-tainted/BgfxRenderer sweep on Linux
Ran the curated clang-tidy over the 22 module files Windows can't parse soundly (bgfx's
compat/mingw/sal.h corrupts the clang std parse) against a Linux/GCC compile DB on VPS142,
finishing the modules/ sweep (PART 1 = the 29 Windows-clean files). 29 findings triaged;
9 real fixes, the rest verified non-bugs left + documented (same policy as PART 1).

THE WIN — a real cross-version portability bug the Linux parse surfaced that Windows built
by luck:
  * SDLMixerBackend.h forward-declared SDL_mixer's INTERNAL struct tag `_Mix_Music`. That
    tag was renamed to `Mix_Music` in newer SDL2_mixer (Debian) → a hard conflicting-typedef
    compile error there; the MinGW toolchain's older SDL2_mixer still used `_Mix_Music` so it
    compiled on Windows. Replaced the fragile (and reserved-identifier) forward-decl with
    `#include <SDL_mixer.h>` — only SDL-aware TUs include this backend header, so no new dep.
    Now compiles on BOTH toolchains (proven: ninja SoundManager_static on Debian + the MinGW
    build + SoundManagerUnit on Windows).

Other real fixes:
  * ~SDLMixerBackend / ~FfmpegCliBackend called a virtual method (shutdown/close) unqualified
    in the destructor -> qualified both (non-virtual dispatch; matches PART 1's ~InputModule).
    The SDLMixer one was HIDDEN behind the parse error until the header fix unblocked it.
  * SDLMixerBackend rounding (int)(v*MAX+0.5) -> std::lround (v clamped >=0 so latent-safe).
  * 5x bugprone-implicit-widening-of-multiplication (FrameAllocator / DebugPass / TilemapPass /
    RHICommandBuffer / BitmapFont) -> target-type casts so the multiply happens in
    size_t/uint32/ptrdiff (no int-overflow-then-widen). All bounded-safe; hardened anyway.
  * BitmapFont::loadBMFont stub -> [[maybe_unused]] on its two params.

Verified NON-bugs LEFT + documented (= PART 1): 7x unused-parameters (IModule/pass override
sigs), 5x enum-size (the one IIO.h IOType core-header spot), 3x int-to-ptr (nativeWindowHandle
is an int-encoded pointer), 2x branch-clone (ShaderManager backend->shader-set select).

Locked by: Windows build + full 153-test ctest (green, no regression) + `ninja
SoundManager_static` on Debian + the clang-tidy RE-RUN showing SDLMixerBackend.cpp at 0
findings and the fixed checks dropped from the histogram (29 -> 18, 0 errors). Recipe +
full triage in docs/design/quality-hardening-handoff.md (Phase 2 UPDATE 2026-07-10).

+++ modules/BgfxRenderer/Frame/FrameAllocator.h
@@ -13,7 +13,7 @@ namespace grove {
 
 class FrameAllocator {
 public:
-    static constexpr size_t DEFAULT_SIZE = 16 * 1024 * 1024; // 16 MB
+    static constexpr size_t DEFAULT_SIZE = static_cast<size_t>(16) * 1024 * 1024; // 16 MB (compute in size_t)
 
     explicit FrameAllocator(size_t size = DEFAULT_SIZE);
     ~FrameAllocator();
